requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 24.11.2
accelerate=1.5.2=pyhd8ed1ab_0
aiohappyeyeballs=2.4.4=py313hca03da5_0
aiohttp=3.11.10=py313h80987f9_0
aiosignal=1.2.0=pyhd3eb1b0_0
arrow-cpp=19.0.0=h0b7d223_2
attrs=24.3.0=py313hca03da5_0
aws-c-auth=0.6.19=h80987f9_0
aws-c-cal=0.5.20=h80987f9_0
aws-c-common=0.8.5=h80987f9_0
aws-c-compression=0.2.16=h80987f9_0
aws-c-event-stream=0.2.15=h313beb8_0
aws-c-http=0.6.25=h80987f9_0
aws-c-io=0.13.10=h80987f9_0
aws-c-mqtt=0.7.13=h80987f9_0
aws-c-s3=0.1.51=h80987f9_0
aws-c-sdkutils=0.1.6=h80987f9_0
aws-checksums=0.1.13=h80987f9_0
aws-crt-cpp=0.18.16=h313beb8_0
aws-sdk-cpp=1.11.212=hdd7fb2f_0
blas=1.0=openblas
boost-cpp=1.82.0=h48ca7d4_2
bottleneck=1.4.2=py313ha35b7ea_0
brotli-python=1.0.9=py313h313beb8_9
bzip2=1.0.8=h80987f9_6
c-ares=1.19.1=h80987f9_0
ca-certificates=2025.2.25=hca03da5_0
certifi=2025.1.31=pyhd8ed1ab_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
datasets=3.3.2=py313hca03da5_0
deprecated=1.2.13=py313hca03da5_0
dill=0.3.8=py313hca03da5_0
expat=2.7.1=h313beb8_0
filelock=3.17.0=py313hca03da5_0
frozenlist=1.5.0=py313h80987f9_0
fsspec=2024.12.0=py313hca03da5_0
gflags=2.2.2=h313beb8_1
glog=0.5.0=h313beb8_1
gmp=6.3.0=h313beb8_0
gmpy2=2.2.1=py313h5c1b81f_0
huggingface_hub=0.29.2=py313hca03da5_0
icu=73.1=h313beb8_0
idna=3.7=py313hca03da5_0
importlib-metadata=8.5.0=py313hca03da5_0
jinja2=3.1.6=py313hca03da5_0
krb5=1.20.1=hf3e1bf2_1
libabseil=20250127.0=cxx17_h313beb8_0
libboost=1.82.0=h0bc93f9_2
libbrotlicommon=1.0.9=h80987f9_9
libbrotlidec=1.0.9=h80987f9_9
libbrotlienc=1.0.9=h80987f9_9
libcurl=8.12.1=hde089ae_0
libcxx=14.0.6=h848a8c0_0
libedit=3.1.20230828=h80987f9_0
libev=4.33=h1a28f6b_1
libevent=2.1.12=h02f6b3c_1
libffi=3.4.4=hca03da5_1
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libgrpc=1.71.0=h62f6fdd_0
libiconv=1.16=h80987f9_3
libmpdec=4.0.0=h80987f9_0
libnghttp2=1.57.0=h62f6fdd_0
libopenblas=0.3.21=h269037a_0
libprotobuf=5.29.3=h9f9f828_0
libre2-11=2024.07.02=h313beb8_0
libssh2=1.11.1=h3e2b118_0
libthrift=0.15.0=h73c2103_2
libtorch=2.5.1=gpu_mps_h7b1bc93_203
libuv=1.48.0=h80987f9_0
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_1
markupsafe=3.0.2=py313h80987f9_0
mpc=1.3.1=h80987f9_0
mpfr=4.2.1=h80987f9_0
mpmath=1.3.0=py313hca03da5_0
multidict=6.1.0=py313h80987f9_0
multiprocess=0.70.15=py313hca03da5_0
ncurses=6.4=h313beb8_0
networkx=3.4.2=py313hca03da5_0
nomkl=3.0=0
numexpr=2.10.1=py313h5d9532f_0
numpy=2.2.4=py313h7c57ca2_0
numpy-base=2.2.4=py313hb98e858_0
openssl=3.5.0=h81ee809_0
opentelemetry-api=1.30.0=py313hca03da5_0
orc=2.1.1=h55d209b_0
packaging=24.2=py313hca03da5_0
pandas=2.2.3=py313hcf29cfe_0
pip=25.0=py313hca03da5_0
propcache=0.3.1=py313h80987f9_0
protobuf=5.29.3=py313h514c7bf_0
psutil=7.0.0=py313h90d716c_0
pyarrow=19.0.0=py313h313beb8_1
pysocks=1.7.1=py313hca03da5_0
python=3.13.2=h4862095_100_cp313
python-dateutil=2.9.0post0=py313hca03da5_2
python-tzdata=2023.3=pyhd3eb1b0_0
python-xxhash=3.5.0=py313h80987f9_0
python_abi=3.13=0_cp313
pytorch=2.5.1=gpu_mps_py313hbd1d365_203
pytz=2024.1=py313hca03da5_0
pyyaml=6.0.2=py313h80987f9_0
re2=2024.07.02=h48ca7d4_0
readline=8.2=h1a28f6b_0
regex=2024.11.6=py313h80987f9_0
requests=2.32.3=py313hca03da5_1
safetensors=0.5.3=py313h00184e3_0
setuptools=72.1.0=py313hca03da5_0
six=1.17.0=py313hca03da5_0
sleef=3.5.1=h80987f9_2
snappy=1.2.1=h313beb8_0
sqlite=3.45.3=h80987f9_0
sympy=1.13.3=py313hca03da5_1
tk=8.6.14=h6ba3021_0
tokenizers=0.21.0=py313h5b23da0_0
tqdm=4.67.1=py313h7eb115d_0
transformers=4.49.0=py313hca03da5_0
typing-extensions=4.12.2=py313hca03da5_0
typing_extensions=4.12.2=py313hca03da5_0
tzdata=2025a=h04d1e81_0
urllib3=2.3.0=py313hca03da5_0
utf8proc=2.6.1=h80987f9_1
wheel=0.45.1=py313hca03da5_0
wrapt=1.17.0=py313h80987f9_0
xxhash=0.8.0=h1a28f6b_3
xz=5.6.4=h80987f9_1
yaml=0.2.5=h1a28f6b_0
yarl=1.18.0=py313h80987f9_0
zipp=3.21.0=py313hca03da5_0
zlib=1.2.13=h18a0788_1
zstd=1.5.6=hfb09047_0
